Write pull and build states after executing steps
Create state files containing contextual information relevant to the
pull and build steps after the corresponding run action is executed.

Note that using the apt cache with the system packages requires
rootdir to be set to "/", otherwise it seems that apt.cache preserves
the state from the previous instance and will keep using the cache
created for stage packages.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>