Implement outdated step verification
Verify if a step is outdated, and update or rerun this step if needed.
This is done by querying the source if the step is PULL, or verifying
previous steps. Since the outdated verification is performed on
persistent (on-disk) data, an `updated` metadata entry was added
to the StateTimestamp state wrapper to mark the part as updated when
the updating action is generated.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>