Create lifecycle execution context
Lifecyle actions should now be executed after a call to execution_start()
to allow the environment to be properly set up. A call to execution_stop()
cleans up the environment after the actions are executed. This allows us
to use more complex environments when processing a parts specification.

Alternatively, action execution can happen inside a context that calls
the guard functions.

Using this new infrastructure, build package installation for all parts
was moved to the execution prologue. This is the behavior we see in
snapcraft, and should marginally improve performance by avoiding
calling the package manager during the execution of the build action
for each part.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>