Create lifecycle execution context
Lifecyle actions must now run in a execution context that calls the
appropriate prologue and epilogue methods to set up and clean the
execution environment. This allows us to use more complex environments
when processing a parts specification.

Using this new infrastructure, build package installation for all parts
was moved to the execution prologue. This is the behavior we see in
snapcraft (and should also work around the slowdown problem seen in
python-apt).

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>