Add API to resolve stage packages and and retrieve state
When installing stage packages directly to the stage tree on execution
prologue we must check if the set of packages to install differs from
the set of packages currently installed (and clean the stage step in
this case). This change surfaces an internal API for package dependency
resolving, and adds a mechanism to retrieve previously stored state
data.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>