Use temporary deb extraction directory under work dir
Extracting deb data to a tempdir under tmp isn't always the best option
because it could be mounted as tmpfs (which doesn't support user extended
attributes). Also if it's a separate filesystem moving files to stage will
involve data copying, so create a temporary directory next to stage (to
be within the same filesystem).

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>