Refactor layering to keep package lists in separate layer
We want to update the package list only once in the layer environment.
The ideal solution would be to export the package list from the host
environment to keep consistency but we need repo support for that
(and also we need to guarantee consistency between host and layer
environments for that to work correctly).

Our current infrastructure allows for easy multi-layering so use these
for now until we investigate how exporting the list could work.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>